Change from Winston to PyPlot for plotting
A larger commit, which constitute the initial move from Winston to PyPlot
for ploting. See testsuites/plot_suites_ODE_fixed.jl for a plotting
implementation entirely with PyPlot. We also explore the performance of
`pwl`'s PR, by including a solvers file which have only the solvers
implemented in the PR.